refactoring to allow learn from annotated plain text and rename HtmlFeatureExtractor to FeatureExtractor
