+++ docs/ref/index.rst
@@ -14,3 +14,5 @@ Reference
    webannotator
    base
    misc
+   token
+   tokenizer
