+++ docs/ref/token.rst
@@ -1,0 +1,7 @@
+Token
+-----
+
+.. automodule:: webstruct.token
+    :members:
+    :undoc-members:
+    :show-inheritance:
