+++ docs/ref/tokenizer.rst
@@ -1,0 +1,7 @@
+Tokenizer
+---------
+
+.. automodule:: webstruct.tokenizer
+    :members:
+    :undoc-members:
+    :show-inheritance:
