+++ webstruct/__init__.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
-from .loaders import WebAnnotatorLoader, GateLoader, HtmlLoader, load_trees
+from .loaders import GateTextLoader, WebAnnotatorLoader, GateLoader, HtmlLoader, load_trees
 from .sequence_encoding import IobEncoder, InputTokenProcessor
-from .feature_extraction import HtmlFeatureExtractor, HtmlTokenizer, HtmlToken
+from .feature_extraction import HtmlFeatureExtractor, FeatureExtractor
+from .tokenizer import HtmlTokenizer, TextTokenizer
+from .token import HtmlToken
 from .wapiti import WapitiCRF, create_wapiti_pipeline
 from .crfsuite import CRFsuiteCRF, CRFsuiteFeatureEncoder, create_crfsuite_pipeline
 from .model import NER
