+++ webstruct/features/__init__.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
+from webstruct.utils import HtmlTokenFeatures
+
 from .block_features import *
 from .token_features import *
 from .data_features import *
 from .global_features import Pattern
-
 
 EXAMPLE_TAGSET = {'ORG', 'PER', 'SUBJ', 'STREET', 'CITY', 'STATE', 'COUNTRY',
                   'ZIPCODE', 'EMAIL', 'TEL', 'FAX', 'SUBJ', 'FUNC', 'HOURS'}
@@ -19,18 +20,21 @@ EXAMPLE_TOKEN_FEATURES = [
     InsideTag('a'),
     InsideTag('strong'),
 
-    token_identity,
-    token_lower,
+    HtmlTokenFeatures(
 
-    token_shape,
-    token_endswith_colon,
-    token_endswith_dot,
-    token_has_copyright,
-    number_pattern,
-    prefixes_and_suffixes,
+        token_identity,
+        token_lower,
 
-    looks_like_year,
-    looks_like_month,
-    looks_like_email,
-    looks_like_street_part,
+        token_shape,
+        token_endswith_colon,
+        token_endswith_dot,
+        token_has_copyright,
+        number_pattern,
+        prefixes_and_suffixes,
+
+        looks_like_year,
+        looks_like_month,
+        looks_like_email,
+        looks_like_street_part,
+    )
 ]
