+++ webstruct/model.py
@@ -7,38 +7,38 @@ import urllib2
 from lxml.html import tostring
 
 from webstruct.loaders import HtmlLoader
-from webstruct.feature_extraction import HtmlTokenizer
+from webstruct.tokenizer import HtmlTokenizer
 from webstruct.sequence_encoding import IobEncoder
-from webstruct.utils import smart_join
+from webstruct.utils import smart_join, get_token_text
 from webstruct.grouping import choose_best_clustering
 from webstruct.webannotator import EntityColors, to_webannotator
 
 
 class NER(object):
     """
-    Class for extracting named entities from HTML.
+    Class for extracting named entities from HTML or plain text.
 
     Initialize it with a trained ``model``. ``model`` must have
-    ``predict`` method that accepts lists of :class:`~.HtmlToken`
+    ``predict`` method that accepts lists of :class:`~.Token`
     sequences and returns lists of predicted IOB2 tags.
     :func:`~.create_wapiti_pipeline` function returns such model.
     """
-    def __init__(self, model, loader=None, html_tokenizer=None,
+    def __init__(self, model, loader=None, tokenizer=None,
                  entity_colors=None):
         self.model = model
         self.loader = loader or HtmlLoader()
-        self.html_tokenizer = html_tokenizer or HtmlTokenizer()
+        self.tokenizer = tokenizer or HtmlTokenizer()
         if entity_colors is None:
             entity_colors = EntityColors()
         self.entity_colors = entity_colors
 
     def extract(self, bytes_data):
         """
-        Extract named entities from binary HTML data ``bytes_data``.
+        Extract named entities from binary data ``bytes_data``.
         Return a list of ``(entity_text, entity_type)`` tuples.
         """
-        html_tokens, tags = self.extract_raw(bytes_data)
-        groups = IobEncoder.group(zip(html_tokens, tags))
+        tokens, tags = self.extract_raw(bytes_data)
+        groups = IobEncoder.group(zip(tokens, tags))
         return _drop_empty(
             (self.build_entity(tokens, tag), tag)
             for (tokens, tag) in groups if tag != 'O'
@@ -49,6 +49,7 @@ class NER(object):
         A convenience wrapper for :meth:`extract` method that downloads
         input data from a remote URL.
         """
+        assert isinstance(self.tokenizer, HtmlTokenizer)
         data = urllib2.urlopen(url).read()
         return self.extract(data)
 
@@ -58,9 +59,9 @@ class NER(object):
         Return a list of ``(html_token, iob2_tag)`` tuples.
         """
         tree = self.loader.loadbytes(bytes_data)
-        html_tokens, _ = self.html_tokenizer.tokenize_single(tree)
-        tags = self.model.predict([html_tokens])[0]
-        return html_tokens, tags
+        tokens, _ = self.tokenizer.tokenize_single(tree)
+        tags = self.model.predict([tokens])[0]
+        return tokens, tags
 
     def extract_groups(self, bytes_data, dont_penalize=None):
         """
@@ -87,7 +88,7 @@ class NER(object):
 
         return entities
 
-    def build_entity(self, html_tokens, tag):
+    def build_entity(self, tokens, tag):
         """
         Join tokens to an entity. Return an entity, as text.
         By default this function uses :func:`webstruct.utils.smart_join`.
@@ -96,14 +97,15 @@ class NER(object):
         and :meth:`extract_groups` results. If this function returns empty
         string or None, entity is dropped.
         """
-        return smart_join(t.token for t in html_tokens)
+        return smart_join(get_token_text(t) for t in tokens)
 
     def annotate(self, bytes_data, pretty_print=False):
         """
         Return annotated HTML data in WebAnnotator format.
         """
         html_tokens, tags = self.extract_raw(bytes_data)
-        tree = self.html_tokenizer.detokenize_single(html_tokens, tags)
+        assert isinstance(self.tokenizer, HtmlTokenizer)
+        tree = self.tokenizer.detokenize_single(html_tokens, tags)
         tree = to_webannotator(tree, self.entity_colors)
         return tostring(tree, pretty_print=pretty_print)
