+++ webstruct/tests/test_pattern_features.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 import unittest
-from webstruct import GateLoader, HtmlTokenizer, HtmlFeatureExtractor
+from webstruct import GateLoader, HtmlTokenizer, FeatureExtractor
 from webstruct.features import token_lower, token_identity, Pattern
-
+from webstruct.utils import HtmlTokenFeatures
 
 class PatternTest(unittest.TestCase):
 
@@ -17,8 +17,8 @@ class PatternTest(unittest.TestCase):
         return html_tokens
 
     def test_pattern(self):
-        featextractor = HtmlFeatureExtractor(
-            token_features = [token_lower, token_identity],
+        featextractor = FeatureExtractor(
+            token_features = [HtmlTokenFeatures(token_lower, token_identity)],
             global_features = [
                 Pattern((-2, 'lower'), (-1, 'lower'))
             ]
@@ -33,8 +33,8 @@ class PatternTest(unittest.TestCase):
         )
 
     def test_pattern_lookups(self):
-        featextractor = HtmlFeatureExtractor(
-            token_features = [token_lower, token_identity],
+        featextractor = FeatureExtractor(
+            token_features = [HtmlTokenFeatures(token_lower, token_identity)],
             global_features=[
                 Pattern((0, 'lower'), (1, 'token'), out_value='OUT'),
             ]
