+++ webstruct/word_tokenizers.py
@@ -2,7 +2,6 @@
 from __future__ import absolute_import
 import re
 
-
 class WordTokenizer(object):
     r"""This tokenizer is copy-pasted version of TreebankWordTokenizer
     that doesn't split on @ and ':' symbols and doesn't split contractions::
